renamed api to client
